+++ source/Roy.Logging/Roy.Logging/Domain/Program/Parameter.cs
@@ -1,4 +1,4 @@
-﻿namespace Roy.Logging.Domain.Application;
+﻿namespace Roy.Logging.Domain.Program;
 
 /// <summary>
 /// Parameter information.
